+++ packages/eslint-config-sds/ts.js
@@ -1,0 +1,10 @@
+const defaultCfg = require('./js');
+
+module.exports = {
+  ...defaultCfg,
+  extends: [
+    'airbnb-base',
+    'airbnb-typescript/base',
+    'plugin:@typescript-eslint/recommended-type-checked',
+  ],
+};