+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 Please have a look at the corresponding [packages](./packages) for a list of recent changes:
 
 - [@sick-design-system/design-tokens](./packages/design-tokens/CHANGELOG.md)
+- [@sick-design-system/eslint-config-sds](./packages/eslint-config-sds/CHANGELOG.md)